chore(core/modules): move resource manager to new sub directory
