+++ apps/core/src/modules/process/manager.ts
@@ -6,7 +6,7 @@ import type {
 import { LogBuffer } from '../buffer/manager';
 import { ConfigManager } from '../config/manager';
 import { wsManager } from '../ws.manager';
-import { resourceManager } from '../resource.manager';
+import { resourceManager } from '../resource/manager';
 
 export class ProcessManager {
 	private state: ServerState = { status: 'stopped', startedAt: null };
