+++ apps/core/src/routes/api/sockets.ts
@@ -11,7 +11,7 @@ import { PermissionManager } from '@fxmanager/shared/utils';
 import { sessionAuth } from '../../middleware/session';
 import { wsManager } from '../../modules/ws.manager';
 import type { AuthedRequest, RouteModule } from '../../types';
-import { resourceManager } from '../../modules/resource.manager';
+import { resourceManager } from '../../modules/resource/manager';
 import { perfManager } from '../../modules/perf/manager';
 
 wsManager.addCheck('console', (admin) => {
